Bump cmake requirements
As of cmake v4.0, compatibility with cmake < 3.5 has been removed and
compatibility with < 3.10 will soon be removed. Ubuntu 18.04, already at
EOL, is shipped with 3.10, I see no reason to support older version of
cmake in 2025.

Add new policy present in cmake >= 3.19. New policy requires file to be
executable, which for doxygen makes sense.

Do not use find_program for doxygen.conf.in, location is hardcoded
anyway. new policy checks for executable flag, which is not present for
the conf.in. find_file is an alternative but since location is fixed by
DOXYGEN_CONF_IN, use EXIST rather than find